Add format option to customize linelists/off-diagonals
Previous versions constrained the line- and off-diagonal-parameters lists to a specific format. This made it difficult to add or remove parameters for future updates, like the addition of first order line-mixing. This commit adds a new option in the header of the linelist and off-diagonal that follows a structure very similar to the 'format' method of strings in python. This will be useful when we add options to ignore parameters not included by the user.