Add Linux and macOS instructions
This commit adds more detailed build instructions for Arch Linux and macOS, and slightly changes the commands to be more general. The inclusion of C++ 23 in CMake was added to prevent the default use of C++ 11 which is missing some features. This was not set as a requirement, as any version above 14 should be able to compile.